コメント追加 Refactoring
